requirements.txt
pyTelegramBotAPI==4.14.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0